Weather_striming/producer/requirements.txt
pandas
kafka-python
hdfs
